move hosted customer routes to own endpiont
